chore(helix-term): extract shared shell command logic
Extract shared shell command handling from `commands.rs` into a dedicated
`commands::shell` module.

The new `shell` module owns `ShellBehavior`, interactive shell prompt
wrappers, synchronous and async shell execution helpers, and shared
selection replacement/insert/append logic.

Static commands and typable commands now import the same shell helpers
instead of keeping the logic in the root commands module.
